+++ dashboard/src/services/restqa/user.js
@@ -1,0 +1,9 @@
+import Http from '../http'
+
+async function getPreferences () {
+  const result = await Http().get('/preferences')
+  return result.data
+}
+export  {
+  getPreferences
+}