+++ dashboard/src/store/modules/restqa/mutations.js
@@ -22,6 +22,9 @@ const mutations = {
   },
   testResult (state, val) {
     state.testResult = val
+  },
+  preferences (state, preferencesObject) {
+    state.preferences = preferencesObject
   }
 }