+++ CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [0.0.30] - 2021-07-09
 
+### Added
+
+* Enable new command to mange the telemetry : `restqa telemetry off` | [#119](https://github.com/restqa/restqa/pull/119)
+
 ### Update
 
 * [Bump commander@8.0.0](https://github.com/tj/commander.js/releases/tag/v8.0.0)